Tests for Meta reset
